updated a packet folder to house serialization, working pyro and comand sending
