+++ Code/CMakeLists.txt
@@ -66,6 +66,7 @@ target_sources(${CMAKE_PROJECT_NAME} PRIVATE
     Core/App/ApplicationLayer/ApplicationHALS/pyro.c
     Core/App/ApplicationLayer/ApplicationHALS/kalman.c
     Core/App/ApplicationLayer/FlightStateMachine/flight_state.c
+    Core/App/ApplicationLayer/ApplicationHALS/packets.c
     
 )
 
@@ -82,7 +83,7 @@ target_include_directories(${CMAKE_PROJECT_NAME} PRIVATE
     Core/App/Outputs/DataLogger
     Core/App/Outputs/LED&buzzer
     Core/App/Outputs/Telemetry
-     Core/App/ApplicationLayer/FlightStateMachine/
+    Core/App/ApplicationLayer/FlightStateMachine/
 
 
 )
