+++ Code/Core/App/AppDrivers/LoRa/LoRa.c
@@ -489,6 +489,7 @@ HAL_StatusTypeDef lora_RX(uint8_t *buff, uint8_t *rx_length, uint8_t max_length,
         result = platform_read(&sx1, REG_IRQ_FLAGS, &buffer, 1);
         if(result != HAL_OK) return HAL_ERROR;
 
+
         if ((HAL_GetTick() - start) >= timeout_ms) {
             return HAL_TIMEOUT;
         }
@@ -499,9 +500,10 @@ HAL_StatusTypeDef lora_RX(uint8_t *buff, uint8_t *rx_length, uint8_t max_length,
     if(buffer & IRQ_PAYLOAD_CRC_ERR_MASK) return HAL_ERROR;
 
     result = platform_read(&sx1, REG_RX_NB_BYTES, &buffer, 1);
+    printf("RX_NB_BYTES = %d\r\n", buffer);
     if(result != HAL_OK) return HAL_ERROR;
 
-    if (buffer >= max_length) return HAL_ERROR;
+    if (buffer > max_length) return HAL_ERROR;
 
     *rx_length = buffer;
